cambio en el tipo de variables que usa el email log, y refactorización, para que coincida con el tipo de dato en la base de datos

+++ email/src/main/java/uni/fis/email/dto/EmailResponse.java
@@ -15,6 +15,6 @@ public class EmailResponse {
     
     private boolean success;
     private String message;
-    private Long logId;
+    private Integer logId;
     private LocalDateTime timestamp;
 }
